Finalizar configuracion de entidades principales
Registro de consultas

+++ API-Veterinaria/API-Veterinaria/Core/DTOs/Mascota/MascotaDTO.cs
@@ -8,7 +8,7 @@ namespace API_Veterinaria.Core.DTOs.Mascota
         public required string Nombre { get; set; }
         public string Especie { get; set; } = string.Empty;
         public string Sexo { get; set; } = string.Empty;
-        public int Edad { get; set; }
+        public decimal Edad { get; set; }
         public decimal? Peso { get; set; }
         public DateTime FechaNacimiento { get; set; }
         public string? Observaciones { get; set; }
